Enable diagonal scaling automatically when it becomes available
Whether 2D and 3D diagonals can be scaled isn't in itself saved as a state, only determined from the bitmask where the diagonal bits are affected when changing the main axes. Once diagonal scaling becomes available when allowing scaling in the base axes it is assumed the user also want to allow scaling in the diagonal.